requirements.txt
boto3>=1.42.43
PyGithub>=2.8.1
python-dotenv>=1.1.0
slack-bolt>=1.21.0
slack-sdk>=3.33.0
